Mostly formatting changes, oh, and yehr, I've got the track submitter to work again!!!
* Changed format of {$this->arrChanges[] = varname} to {$this->arrChanges[varname] = true} throughout (as it should have been, damn sloppy coding there, Jon)
* Made GeneralFunctions::getFileLengthString() refer to the actual, you know, function in the getID3 library, rather than re-writing my own from scratch.
* Fixed a layout issue in GenericObject::write() where the SQL might add an extra , if the column name wasn't in the list of acceptable columns. Oops!
* On a related note, added "timeLength" to the TrackObject list of acceptable columns. Oops! Made sure we created it in the NewTrackObject class
* Use a session variable to pass back the intTrackID to the HTML class (need to go back and check the API class) from the NewTrackObject class.
* Stopped cloning objects by re-polling them from the database in the HTML class when we've got a processingID (hint, if you've got that, you've got the record already dimwit)
* Removed some fields we don't need from the Jamendo poller.
* In the TrackObject, when we're replacing the fileSource value with a new one - let the rest of the app handle where to store it, just store the file name, otherwise other functions get confused.

+++ CLASSES/class_RemoteSources.php
@@ -301,7 +301,7 @@ class RemoteSources extends GenericObject
     {
         if ( ! $this->inJson($this->strTrackName, $strTrackName)) {
             $this->strTrackName = $this->addJson($this->strTrackName, $strTrackName);
-            $this->arrChanges[] = 'strTrackName';
+            $this->arrChanges['strTrackName'] = true;
         }
     }
 
@@ -316,7 +316,7 @@ class RemoteSources extends GenericObject
     {
         if ($this->preferredJson($this->strTrackName) != $strTrackName) {
             $this->strTrackName = $this->addJson($this->strTrackName, $strTrackName, true);
-            $this->arrChanges[] = 'strTrackName';
+            $this->arrChanges['strTrackName'] = true;
         }
     }
 
@@ -330,7 +330,7 @@ class RemoteSources extends GenericObject
     function del_strTrackName($strTrackName = "")
     {
         $this->strTrackName = $this->delJson($this->strTrackName, $strTrackName);
-        $this->arrChanges[] = 'strTrackName';
+        $this->arrChanges['strTrackName'] = true;
     }
 
     /**
@@ -344,7 +344,7 @@ class RemoteSources extends GenericObject
     {
         if ($this->strTrackNameSounds != $strTrackNameSounds) {
             $this->strTrackNameSounds = $strTrackNameSounds;
-            $this->arrChanges[] = 'strTrackNameSounds';
+            $this->arrChanges['strTrackNameSounds'] = true;
         }
     }
 
@@ -359,7 +359,7 @@ class RemoteSources extends GenericObject
     {
         if ( ! $this->inJson($this->strTrackUrl, $strTrackUrl)) {
             $this->strTrackUrl = $this->addJson($this->strTrackUrl, $strTrackUrl);
-            $this->arrChanges[] = 'strTrackUrl';
+            $this->arrChanges['strTrackUrl'] = true;
         }
     }
 
@@ -374,7 +374,7 @@ class RemoteSources extends GenericObject
     {
         if ($this->preferredJson($this->strTrackUrl) != $strTrackUrl) {
             $this->strTrackUrl = $this->addJson($this->strTrackUrl, $strTrackUrl, true);
-            $this->arrChanges[] = 'strTrackUrl';
+            $this->arrChanges['strTrackUrl'] = true;
         }
     }
 
@@ -388,7 +388,7 @@ class RemoteSources extends GenericObject
     function del_strTrackUrl($strTrackUrl = "")
     {
         $this->strTrackUrl = $this->delJson($this->strTrackUrl, $strTrackUrl);
-        $this->arrChanges[] = 'strTrackUrl';
+        $this->arrChanges['strTrackUrl'] = true;
     }
 
     /**
@@ -708,7 +708,7 @@ class RemoteSources extends GenericObject
             $this->isNSFW,
             $this->fileName
         );
-        if ($this->fileUrl != '' and $track != false) {
+        if ($this->fileUrl != '' and $track != false and file_exists($this->fileName)) {
             unlink($this->fileName);
         }
         if ($track != false) {
